+++ packages/server/src/middlewares/extendReqRes.js
@@ -1,6 +1,6 @@
 import forEach from 'lodash/forEach';
 
-export default ctx => (req, res, next) => {
+export default (ctx) => (req, res, next) => {
   if (ctx.responses) {
     forEach(ctx.responses, (val, key) => {
       res[key] = val.bind(res);