+++ packages/server/src/middlewares/parseToken.js
@@ -1,4 +1,4 @@
-export default ctx => (req, res, next) => {
+export default (ctx) => (req, res, next) => {
   const token = ctx.helpers.getToken(req);
   req.token = token;
   next();