+++ packages/server/src/middlewares/reqData.js
@@ -1,7 +1,7 @@
 import get from 'lodash/get';
 import getReqData from '../utils/getReqData';
 
-export default ctx => {
+export default (ctx) => {
   if (!get(ctx, 'serverConfig.middlewares.reqData')) return null;
   return (req, res, next) => {
     req.data = getReqData(req, get(ctx, 'config.middlewares.reqData.parse'));